add useHistory in home and useParams hook in map.js, in order to push everything the user types inside the params of the other page
